Check and warn if a Cluster is in cluster_maintenance
As a first step in preventing simultaneously running cluster upgrades,
the scheduling_policy of the cluster is checked.  If the policy is
"cluster_maintenance", an extra warning dialog is displayed to the
user.  The user may select to continue to the cluster upgrade despite
the warning.

A new spinner dialog was also added since we want the continue
confirmation dialog to be displayed before the wizard is displayed.

Note: Currently the spinner dialog cannot be canceled or closed. It
will remain open until the `ClusterUpgradeDataProvider` finishes
loading the data.  A future enhancement would be to allow the user to
close the dialog and have that cancel the data fetches in the data
provider.

Change-Id: I3cda2216b6fd0d1cdbc821146ca76a59cd424d38
Bug-Url: https://bugzilla.redhat.com/1687645
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>