vm-migrate: change handling of empty API response
Engine can send back an empty JSON object { } when
there are no hosts to migrate to, including the current one
the VM is on. In other words, sometimes that current one is
included, and sometimes it's not. Handle both cases the same,
i.e. no longer error when none are returned.

Change-Id: I692ce4c19871841e8319478d9a1299428f51ed41
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1670701
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>