vm-migrate: Add checkbox to select if other vms in affinity should be migrated too
User can choose to migrate not only the selected VMs but also
all other VMs that are in positive enforcing affinity with
the VMs.

The checkbox has to be selected also in case, that the selected VMs
are in affinity together.

Another change is that in case there are no valid hosts
for migration, the dialog is not closed, so the user can
select the checkbox.

Screenshot of the new migration dialog: https://imgur.com/a/6FDP5NR

Change-Id: I994889de662da7a06d56abc3825c07ca4f045d82
Bug-Url: https://bugzilla.redhat.com/1651406
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>