Dashboard: handle missing values from fetchData()
Previously, if the dashboard points to an ovirt-engine that doesn't
generate, for any reason, a particular data sub-category, the transform
function will fail.  This causes the dashboard to fail rendering.

This patch adds sub-category existence checks.  If the sub-category
is missing, a default empty value is used.  This will allow the
transform to complete and the dashboard can then render empty state.

Change-Id: I5a202b5888386dcff7197aee6d9ea88fc97c6030
Bug-Url: https://bugzilla.redhat.com/1713967
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>