Fix ClusterUpgradeWizard error going to the review step
With webpack 4, tree shaking is turned on by default in "production"
mode.  Due to a bug on patternfly-react's `Card` component, the
tree shaking step removes the `Card/index.js` file and therefore
removes the code that adds the various sub-components to the `Card`
component itself.  When the `UpgradeReview` component attempted to
render a <Card>, the <Card.Title> and <Card.Body> components are
`undefined` and the react app errors out.  I logged issue [1]
against patternfly-react.  The issue does not happen in "development"
mode since tree shaking in not turned on by default.

To work around that issue, the upgrade review component now uses
each sub-component via direct import.

[1] - https://github.com/patternfly/patternfly-react/issues/2931

Change-Id: I7063d3b6db9c5d0297090af5cadd51b3eb8c949e
Bug-Url: https://bugzilla.redhat.com/1750669
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>