update i18n normalize and unit-tests, bump jest version
Since the project uses gettext when handling translations with
Zanata, when multiple keys have the same base English translation,
only one of those keys will actually be translated.  Gettext does
NOT use a message id as the translation key, it uses the base English
text.  The normalize script has been updated to handle this scenario
by finding the translated key and copying the translation back to the
other keys.  This is necessary so all of the keys that share English
text will show up as translated when you run `yarn intl:report`.

Change-Id: Ief768455fa49ba97e4cdcbb250b3a6014b36829b
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>