Add copy host networks functionality
Enable selecting a source host and opening a dialog to select a target
host out of a set of eligible hosts. The networks of the source host
will be copied to the target host using existing engine functionality,
via a REST API call.

A button is added to the main hosts list page and to the main host page.
On the hosts lists page it should be enabled only when a single host is
selected. On the main host page it should always be enabled.

Change-Id: If424ff60333b38f346883edb3bb85487a90bba41
Bug-Url: https://bugzilla.redhat.com/1768937
Signed-off-by: Eitan Raviv <eraviv@redhat.com>