Add 'Display on' to vGPU dialog
The vGPU dialog lacks the possibility to set the 'nodisplay' property to
the vm custom properties. To avoid the negation in the dialog, it shows
a checkbox with label "Display on" that is by default selected. If it is
unchecked, the 'nodisplay' item is added at the beginning of the
custom properties field.

Change-Id: I9ccaf47dab3845bec059b2537a78656b1a970ba0
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>