Update Zanata tools to use the maven plugin
Since the zanata-cli rpm is old on F30 and may not exist in more
recent Fedora or CentOS releases, the Zanata client used has been
updated.  Now, the Zanata maven plugin is being used.  Java 1.8 is
still a requirement to run the plugin, just like the zanata-cli.

Renamed the yarn script targets to `intl:push` and `intl:pull`
to align with ovirt-web-ui.  README.adoc updated to reflect the
target name changes.

Change-Id: I0a197d33b50676378365dc37915da1723287b354
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>