vGPU dialog: add more instances of a card
This patch allows to request more instances of the same card. The
currently used Patternfly version does not allow to select how the
table row action is displayed, it is always displays as kebab. This
should change in the next Patternfly version and it might be worth to
update the vGPU dialog for better usability.

Change-Id: If4743fb771524a3f9ca0076fad670866bbc0e8e8
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1860646