Improve logging and error handling
Up until now, there was no error  handling
if the engine rest API call failed. This patch
checks if the response is OK and if not, creates
an appropriate error message and throws an exception.

It also introduces a global error handler that calls
the webadmin's reportFatalError method to
log the uncaught error in the remote ui.log file and
show an Uncaught error occurred dialog in the webadmin.

Change-Id: Ic9a3fe2303311fc3efed36a98690c1795610db75
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1771925