vGPU: Fix the max instances
The max instances for the vGPU type should be calculated
as the maximum of the sums of max instances per host. E.g.

Host A has 2 vGPU cards, each with max instances 4
Host B has 4 vGPU cards, each with max instances 4

That means that the sum of all max instances on all host's cards
is 8 for Host A and 16 for Host B. The vGPU type's max instances is
16.

Change-Id: I7b00b073faafad7b5251549164d12dc9090e6afa
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1980331