cluster upgrade: use correlation id for ansible playbooks
Create a random 10 hex-digit string to use as a correlation id
when starting the cluster upgrade playbook.  The correlation id
will be used by all of the progress events and by the other tasks
run by the ansible role.

Once the playbook has started, push the user to the events grid
with a filter set for the correlation id.  This will show the progress
of the upgrade as it happens.  Any failures encountered by the role
should also be visible.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2040474
Change-Id: Ibce8d71da169634d3ab5578ac39b80ac2846a533
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>