Add link to Monitoring Portal on Dashboard
If the URL for the engine's grafana instance is available from
the `dashboard_data` fetch, render the link in the upper right
hand side of the dashboard.  This allows the user to click directly
to the monitoring portal without jumping back to the welcome page
first.

If no URL is provided from the `dashboard_data` fetch, no link is
rendered.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1991482
Require ovirt-engine PR: https://github.com/oVirt/ovirt-engine/pull/85