cluster upgrade: push i18n directly to where it is used
Cluster upgrade was written to be able to be potentially used outside
of ovirt-engine-ui-extensions.  This support was primarly to keep i18n
use extenal to the components.  The support is no longer necessary
and complicates the code as compared to the other modals.  By pushing
the i18n use directly to the components, everything is simpler and
more consistent.