cluster upgrade: Add a track progress step
After the review step, add a `isFinishedStep` track progress step
to allow the user to monitor the progress of the upgrade.  On entering
the step, the wizard navigation and footer is hidden.  `progress`
data is looked up and displayed by the tracker.  The tracking is done
by querying events with the correlation id used to start the upgrade
process.

Summary of changes:

  - Extracted data and rest related functions to `data.js`

  - `ClusterUpgradeModal` uses explicit components for each loading
    variation

  - Updated `ClusterUpgradeDataProvider` to use the `useDataProvider()`
    hook to handle fetching.  This makes handling tracking updates much
    easier to manage.

  - Add `TrackProgress` component to fetch and display progress
    information and related events in decending event id order

  - Restyle the PF4 "Wizard in a Modal" to look like a PF3 Wizard
    modal box.  The styles will match the other modals.