build: support HEAD with multiple tags (#44)
In the case of a HEAD commit having multiple tags, the existing
build script would fail to recognize it as a tagged commit and
build as a snapshot build.  The simple fix to add some quotes fixes
this issue.  Now a commit with any number of tags will produce a
release build.