vm-manage-gpu: Update mdevs (#45)
* Modify vGPU dialog to use mdev devices REST API

Previously, the vGPU was configured using custom properties. This
has been recently changed on the engine to use a special mdev
device type. This patch reflects that change.

Bug-Url: https://bugzilla.redhat.com/1987121

* Add driver parameters to vGPU dialog

The driver parameters are a free string that is passed
without any checks to the host's driver. It is up to the
user to check the driver's documentation for proper format.

Bug-Url: https://bugzilla.redhat.com/1987121