cluster upgrade: migrate the wizard to PF4, add progress view (#37)
* cluster upgrade: push i18n directly to where it is used

Cluster upgrade was written to be able to be potentially used outside
of ovirt-engine-ui-extensions.  This support was primarly to keep i18n
use extenal to the components.  The support is no longer necessary
and complicates the code as compared to the other modals.  By pushing
the i18n use directly to the components, everything is simpler and
more consistent.

* Fix react lint warning on PluginApiModal

* cluster upgrade: upgrade to PF4

The cluster upgrade wizard has been upgrade following the current
PatternFly 4 standards.  Techniques and styles used are consistent
with the other PF4 modals.  The select hosts table is now using
the `TableComposable` style instead of the legacy component.  This
helps keep the table complexity low.

The old `ClusterUpgradeWizard` was broken up to `ClusterUpgradeModal`
and the new `ClusterUpgradeWizard` to separate the concerns of the
loading spinner and maintenance policy confirm checks from the wizard
itself.

Remove old PF3 components / helpers that are no longer needed:
  - `BaseFormGroup`
  - `FieldLevelHelp` (the PF3 version)
  - `WizardPattern`
  - `MessageDialog`
  - `SpinnerDialog`

* cluster upgrade: Add a track progress step

After the review step, add a `isFinishedStep` track progress step
to allow the user to monitor the progress of the upgrade.  On entering
the step, the wizard navigation and footer is hidden.  `progress`
data is looked up and displayed by the tracker.  The tracking is done
by querying events with the correlation id used to start the upgrade
process.

Summary of changes:

  - Extracted data and rest related functions to `data.js`

  - `ClusterUpgradeModal` uses explicit components for each loading
    variation

  - Updated `ClusterUpgradeDataProvider` to use the `useDataProvider()`
    hook to handle fetching.  This makes handling tracking updates much
    easier to manage.

  - Add `TrackProgress` component to fetch and display progress
    information and related events in decending event id order

  - Restyle the PF4 "Wizard in a Modal" to look like a PF3 Wizard
    modal box.  The styles will match the other modals.