github actions: always upgrade nodejs-modules (#47)
Since we use the ovirt buildcontainer[1] with ovirt-engine-nodejs-modules
already installed, there is a gap in time from when we update
nodejs-modules and when it becomes available in the container.  The container
builds at least once a day.

To facilitate our use of nodejs-modules and allow us to use pre-seed
updates as soon as they are available in copr, the github actions are
updated to always `dnf upgrade` the package.  If no upgrades are available,
the operation will take very little time.

[1] - https://github.com/oVirt/buildcontainer