test: fix error in testing based on node version
When running yarn test locally a different result in contrast to the CI was generated based on the node version installed.
This due to updates to the Intl DatetimeFormatter that generates different outcomes.
Usage of an if statement due to centos9 having a max node version of 16.20 so support is still needed.

Signed-off-by: Jasper Berton <jasper.berton@team.blue>