ui: ability to search for targets on storage connections pop-up screen
Similar to the manage storage domain screen, one can now search for storage connections based on port, ip and host.
Disables searching and adding when storage domain is not in maintenance due to this not being supported at the moment.
Only show hosts that are up in dropdown menu so non usable hosts do not show up.

Signed-off-by: Jasper Berton <jasper.berton@team.blue>